+++ CHANGELOG.md
@@ -1,5 +1,86 @@
 # CHANGELOG
 
+## 1.2.24
+Build date: Aug 17, 2024.
+
+Features & Update:
+- Update balance calculation for nomination pool runtime update (#3443)
+- Do not allow send to empty account (Native token balance = 0) (#2783)
+- Add validate for Solochain when receive, transfer with Generic ledger account (#3464)
+- Update chainlist (#3451)
+  - Add support for Blast Mainnet
+  - Add in-app staking support for DENTX
+
+Bugs fixed:
+- Fix Input overflow width issue (#3441)
+- Fix bug Do not show Signature popup in case not enough balance to cover gas fee (#2501)
+
+
+## 1.2.24
+Build date: Aug 09, 2024.
+
+Features & Update:
+- Allow to use Migration Polkadot App to attach Ledger account (#3307, #3402)
+- Hide direct api usage of polkadot/js (#3308)
+- Improve performance upon showing Marketing Campaign (#3414)
+- Update chainlist (#3425)
+  - Add support for Pendulum assets (ASTR, vDOT, BNC)
+  - Add support for Exosama network
+
+
+## 1.2.23
+Build date: Aug 03, 2024.
+
+Features & Update:
+- Update chain-list (#3403)
+  - Add support for ROUTE token
+  - Add support for Aleph Zero EVM (mainnet)
+  - Update price-id for AZERO EVM
+
+
+## 1.2.22
+Build date: Jul 31, 2024.
+
+Features & Update:
+- Update chain-list (#3385)
+- Integrate Avail Ledger app (#2982)
+
+Bugs fixed:
+- Fix bug Show incorrect withdrawal information of the Bifrost liquid staking when all accounts mode (#3327)
+- Fix bug do not show root screen after remove account (#3148)
+- Fix bug Show blank screen when attach account (#3054)
+
+
+## 1.2.21
+Build date: Jul 24, 2024.
+
+Bugs fixed:
+Fix a few bugs for Avail (#3378)
+  - Add AVAIL token on Ethereum
+  - Integrate Avail OG NFT
+  - Setup default nomination pool online
+  - Update Avail staking APY
+  - Fix bug when click "View details" on the Sign transaction from dApp
+
+## 1.2.20
+Build date: Jul 24, 2024.
+
+Features & Update:
+- Update chain-list for Avail
+
+
+## 1.2.19
+Build date: Jul 23, 2024.
+
+Features & Update:
+- Update chain-list for Avail
+
+## 1.2.18
+Build date: Jul 23, 2024.
+
+Features & Update:
+- Update chain-list for Avail (#3369)
+
 ## 1.2.17
 Build date: Jul 22, 2024.